requirements.txt
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
great-expectations==0.18.21
dbt-postgres==1.8.2
